+++ glib/gmath.cpp
@@ -17,8 +17,8 @@ namespace glib {
 template <class T> using iType = iterator<T>;
 
 template <class T> std::ostream& operator<< (std::ostream& out, const Matrix<T>& rhs) {
-    for (size_t i = 0; i < rhs.n; ++i) {
-        for (size_t j = 0; j < rhs.m; ++j) {
+    for (gint i = 0; i < rhs.n; ++i) {
+        for (gint j = 0; j < rhs.m; ++j) {
             out << rhs.at(i, j) << rhs.ws;
         }
         out << '\n';
@@ -27,7 +27,7 @@ template <class T> std::ostream& operator<< (std::ostream& out, const Matrix<T>&
 }
 
 template <class T> std::istream& operator>> (std::istream& in, Matrix<T>& rhs) {
-    for (size_t i = 0; i < rhs.length; ++i) {
+    for (gint i = 0; i < rhs.length; ++i) {
         in >> rhs.elem[i];
     }
     return in;
@@ -35,7 +35,7 @@ template <class T> std::istream& operator>> (std::istream& in, Matrix<T>& rhs) {
 
 template <class T> Matrix<T> operator+ (const T& lhs, const Matrix<T>& rhs) {
     Matrix<T> res(rhs.n, rhs.m);
-    for (size_t i = 0; i < rhs.length; ++i) {
+    for (gint i = 0; i < rhs.length; ++i) {
         res.elem[i] = lhs + rhs.elem[i];
     }
     return res;
@@ -48,7 +48,7 @@ template <class T> Matrix<T> operator+ (const Matrix<T>& lhs, const T& rhs) {
 template <class T> Matrix<T> operator+ (const Matrix<T>& lhs, const Matrix<T>& rhs) {
     if (lhs.n == rhs.n && lhs.m == rhs.m) {
         Matrix<T> res(lhs.n, rhs.m);
-        for (size_t i = 0; i < lhs.length; ++i) {
+        for (gint i = 0; i < lhs.length; ++i) {
             res[i] = lhs[i] + rhs[i];
         }
         return lhs;
@@ -58,7 +58,7 @@ template <class T> Matrix<T> operator+ (const Matrix<T>& lhs, const Matrix<T>& r
 
 template <class T> Matrix<T> operator- (const T& lhs, const Matrix<T>& rhs) {
     Matrix<T> res(rhs.n, rhs.m);
-    for (size_t i = 0; i < rhs.length; ++i) {
+    for (gint i = 0; i < rhs.length; ++i) {
         res.elem[i] = lhs - rhs.elem[i];
     }
     return res;
@@ -71,7 +71,7 @@ template <class T> Matrix<T> operator- (const Matrix<T>& lhs, const T& rhs) {
 template <class T> Matrix<T> operator- (const Matrix<T>& lhs, const Matrix<T>& rhs) {
     if (lhs.n == rhs.n && lhs.m == rhs.m) {
         Matrix<T> res(lhs.n, rhs.m);
-        for (size_t i = 0; i < lhs.length; ++i) {
+        for (gint i = 0; i < lhs.length; ++i) {
             res[i] = lhs[i] - rhs[i];
         }
         return lhs;
@@ -81,7 +81,7 @@ template <class T> Matrix<T> operator- (const Matrix<T>& lhs, const Matrix<T>& r
 
 template <class T> Matrix<T> operator* (const T& lhs, const Matrix<T>& rhs) {
     Matrix<T> res(rhs.n, rhs.m);
-    for (size_t i = 0; i < rhs.length; ++i) {
+    for (gint i = 0; i < rhs.length; ++i) {
         res.elem[i] = lhs * rhs.elem[i];
     }
     return res;
@@ -95,10 +95,10 @@ template <class T> Matrix<T> operator* (const Matrix<T>& lhs, const Matrix<T>& r
     if (lhs.m == rhs.n) {
         T sum;
         Matrix<T> res(lhs.n, rhs.m);
-        for (size_t i = 0; i < res.n; ++i) {
-            for (size_t j = 0; j < res.m; ++j) {
+        for (gint i = 0; i < res.n; ++i) {
+            for (gint j = 0; j < res.m; ++j) {
                 sum = 0;
-                for (size_t k = 0; k < lhs.m; ++k) {
+                for (gint k = 0; k < lhs.m; ++k) {
                     sum += lhs.at(i, k) * rhs.at(k, j);
                 }
                 res.at(i, j) = sum;
@@ -111,7 +111,7 @@ template <class T> Matrix<T> operator* (const Matrix<T>& lhs, const Matrix<T>& r
 
 template <class T> Matrix<T> epow(const Matrix<T>& lhs, double exp) {
     Matrix<T> res(lhs.n, lhs.m);
-    for (size_t i = 0; i < lhs.length; ++i) {
+    for (gint i = 0; i < lhs.length; ++i) {
         res.elem[i] = std::pow(lhs.elem[i], exp);
     }
     return res;
@@ -119,7 +119,7 @@ template <class T> Matrix<T> epow(const Matrix<T>& lhs, double exp) {
 
 template <class T> Matrix<T> elog(const Matrix<T>& lhs) {
     Matrix<T> res(lhs.n, lhs.m);
-    for (size_t i = 0; i < lhs.length; ++i) {
+    for (gint i = 0; i < lhs.length; ++i) {
         res.elem[i] = std::log10(lhs.elem[i]);
     }
     return res;
@@ -127,7 +127,7 @@ template <class T> Matrix<T> elog(const Matrix<T>& lhs) {
 template <class T> Matrix<T> edot(const Matrix<T>& lhs, const Matrix<T>& rhs) {
     if (lhs.n == rhs.n && lhs.m == rhs.m) {
         Matrix<T> res(lhs.n, lhs.m);
-        for (size_t i = 0; i < lhs.length; ++i) {
+        for (gint i = 0; i < lhs.length; ++i) {
             res.elem[i] = lhs.elem[i] * rhs.elem[i];
         }
         return res;
@@ -145,7 +145,7 @@ template <class T> Matrix<T>::Matrix(const Matrix &rhs) :
     idx(pDim{rhs.n, rhs.m})
 {
     resize(rhs.n, rhs.m);
-    for (size_t i = 0; i < length; ++i) {
+    for (gint i = 0; i < length; ++i) {
         elem[i] = rhs.elem[i];
     }
 }
@@ -168,7 +168,7 @@ template <class T> Matrix<T>& Matrix<T>::operator=(const Matrix& rhs) {
     length = n * m;
 
     elem = new T[length];
-    for (size_t i = 0; i < length; ++i) {
+    for (gint i = 0; i < length; ++i) {
         elem[i] = rhs.elem[i];
     }
     return *this;
@@ -233,7 +233,7 @@ template<class T> void Matrix<T>::importIDX(const char* filename) {
     // Reading data
 
     unsigned char byte;
-    size_t i = 0;
+    gint i = 0;
 
     while (inf.read(reinterpret_cast<char*>(&byte), datasize) && i < items) {
         elem[i] = byte;
@@ -245,7 +245,7 @@ template<class T> void Matrix<T>::exportTXT(const char* file) {
     std::ofstream outf(file, std::ios::out);
     std::stringstream ss;
     std::ios::sync_with_stdio(false);
-    for (size_t i = 0; i < n; ++i) {
+    for (gint i = 0; i < n; ++i) {
         outf << getRow(i);
     }
     outf.close();
@@ -267,8 +267,8 @@ template<class T> void Matrix<T>::resize(gint n, gint m) {
 // Transpose
 template <class T> Matrix<T> Matrix<T>::tr() {
     Matrix<T> res(m, n);
-    for (size_t i = 0; i < res.n; ++i) {
-        for (size_t j = 0; j < res.m; ++j) {
+    for (gint i = 0; i < res.n; ++i) {
+        for (gint j = 0; j < res.m; ++j) {
             res.at(i, j) = this->at(j, i);
         }
     }
@@ -277,7 +277,7 @@ template <class T> Matrix<T> Matrix<T>::tr() {
 
 template <class T> void Matrix<T>::set(const Matrix<T>& in) {
     if (in.n == n && in.m == m) {
-        for (size_t i = 0; i < length; ++i) {
+        for (gint i = 0; i < length; ++i) {
             elem[i] = in.elem[i];
         }
     }
@@ -302,7 +302,7 @@ template <class T> gint Matrix<T>::getM()                           { return m; 
 
 template <class T> Matrix<T> Matrix<T>::getRow(gint r) {
     Matrix<T> row(1, m);
-    for (size_t i = 0; i < m; ++i) {
+    for (gint i = 0; i < m; ++i) {
         row.elem[i] = elem[idx.at(r, i)];
     }
     //memcpy(&row.elem, &elem[idx.at(r, 0)], sizeof row.elem[0] * m);
@@ -321,7 +321,7 @@ inline double estDerivative(double x)       { return (x + epsilon - (x - epsilon
 template <class T> Matrix<T> sigmoid(Matrix<T>& x) {
     gint length = x.getLength();
     Matrix<T> res(x.getN(), x.getM());
-    for (size_t i = 0; i < length; i++)    {
+    for (gint i = 0; i < length; i++)    {
         res[i] = sigmoid(x[i]);
     }
     return res;
